requirement.txt
asgiref==3.5.2
autopep8==1.7.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
lxml==4.9.1
numpy==1.23.3
panda==0.3.1
pandas==1.5.0
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.4
urllib3==1.26.12
